Added Menu class and refactored QuizHandler. Added 10 more Git questions.
